Ajustes de UI no Header e CollectionView (com TableView), Mecanismo do Navegação e nomenclatura do protocolo Navigable
