+++ YogaZone/YogaZone/Home/Menu/HeaderCell/HeaderCell.swift
@@ -1,0 +1,37 @@
+//
+//  HeaderCell.swift
+//  YogaZone
+//
+//  Created by Philippe Muniz Gomes on 23/03/22.
+//
+
+import UIKit
+
+class HeaderCell: UITableViewCell {
+    @IBOutlet weak var greetingLabel: UILabel!
+    @IBOutlet weak var usernameLabel: UILabel!
+    @IBOutlet weak var profileImageView: UIImageView!
+    
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        
+    }
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        self.profileImageView.layer.cornerRadius =  30
+        self.profileImageView.clipsToBounds = true
+        self.profileImageView.contentMode = .scaleAspectFill
+        self.profileImageView.image = #imageLiteral(resourceName: "profile-2")
+    }
+}
+
+// MARK: Static properties & methods
+extension HeaderCell {
+    static let identifier: String = "HeaderCell"
+    
+    static func getNib() -> UINib {
+        return UINib(nibName: self.identifier, bundle: nil)
+    }
+}
+
